Fix weird line artifact for triangle rasterization
This bug happened because screen coordinates were originally supposed to be ints.
The recent refactor made it so that the screen coordinates now include a z axis that's
a real number. So now we are dealing with Vec3f. So I just casted it to an i32 to drop
the decimal and cast it back to a float for Vec3f.

I still have the weird missing triangle bug...